backend/service-visual-capture/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
opencv-python-headless>=4.9.0
onvif-zeep>=0.2.12
Pillow>=10.0.0
